Split rtm api caching into separate file
